+++ src/components/grouped-character/grouped-character-test.stories.tsx
@@ -6,11 +6,15 @@ import {
   getCommonTextboxArgsWithSpecialCaracters,
   CommonTextboxArgs,
 } from "../textbox/textbox-test.stories";
-import GroupedCharacter, { CustomEvent } from "./grouped-character.component";
+import GroupedCharacter, {
+  CustomEvent,
+  GroupedCharacterProps,
+} from "./grouped-character.component";
 import CarbonProvider from "../carbon-provider/carbon-provider.component";
 
 export default {
   title: "GroupedCharacter/Test",
+  includeStories: "Default",
   parameters: {
     info: { disable: true },
     chromatic: {
@@ -76,4 +80,31 @@ NewValidation.args = {
   groups: [2, 2, 4],
   separator: "-",
   ...getCommonTextboxArgs(),
+};
+
+export const GroupedCharacterComponent = ({
+  onChange,
+  groups,
+  separator,
+  ...props
+}: GroupedCharacterProps) => {
+  const [state, setState] = React.useState("");
+
+  const setValue = (ev: CustomEvent) => {
+    setState(ev.target.value.rawValue);
+    if (onChange) {
+      onChange(ev);
+    }
+  };
+
+  return (
+    <GroupedCharacter
+      label="GroupedCharacter"
+      value={state}
+      onChange={setValue}
+      groups={groups || [2, 2, 3]}
+      separator={separator || "-"}
+      {...props}
+    />
+  );
 };