+++ src/components/inline-inputs/inline-inputs.component.tsx
@@ -1,4 +1,4 @@
-import React, { useRef } from "react";
+import React from "react";
 import { MarginProps } from "styled-system";
 import Label from "../../__internal__/label";
 import StyledInlineInputs, {
@@ -8,13 +8,8 @@ import StyledInlineInputs, {
   StyledInlineInputsProps,
 } from "./inline-inputs.style";
 import FormSpacingProvider from "../../__internal__/form-spacing-provider";
-import createGuid from "../../__internal__/utils/helpers/guid";
 import useIsAboveBreakpoint from "../../hooks/__internal__/useIsAboveBreakpoint";
 import useFormSpacing from "../../hooks/__internal__/useFormSpacing";
-
-interface InlineInputsContextProps {
-  ariaLabelledBy?: string;
-}
 
 type GutterOptions =
   | "none"
@@ -40,26 +35,21 @@ export interface InlineInputsProps
   htmlFor?: string;
   /** Defines the label text for the heading. */
   label?: string;
+  /**
+   * Custom label id, could be used in combination with aria-labelledby prop of each input,
+   * to make them accesible for screen readers.
+   */
+  labelId?: string;
   /** Flag to configure component as mandatory. */
   required?: boolean;
 }
 
-export const InlineInputsContext: React.Context<InlineInputsContextProps> = React.createContext(
-  {}
-);
-
-const columnWrapper = (
-  children: React.ReactNode,
-  gutter: GutterOptions,
-  labelId?: string
-) => {
+const columnWrapper = (children: React.ReactNode, gutter: GutterOptions) => {
   return React.Children.map(children, (input) => {
     return (
-      <InlineInputsContext.Provider value={{ ariaLabelledBy: labelId }}>
-        <StyledInlineInput gutter={gutter} data-element="inline-input">
-          {input}
-        </StyledInlineInput>
-      </InlineInputsContext.Provider>
+      <StyledInlineInput gutter={gutter} data-element="inline-input">
+        {input}
+      </StyledInlineInput>
     );
   });
 };
@@ -67,6 +57,7 @@ const columnWrapper = (
 const InlineInputs = ({
   adaptiveLabelBreakpoint,
   label,
+  labelId,
   htmlFor,
   children = null,
   className = "",
@@ -77,7 +68,6 @@ const InlineInputs = ({
   required,
   ...rest
 }: InlineInputsProps) => {
-  const labelId = useRef(createGuid());
   const largeScreen = useIsAboveBreakpoint(adaptiveLabelBreakpoint);
   let inlineLabel: boolean | undefined = labelInline;
   if (adaptiveLabelBreakpoint) {
@@ -89,7 +79,7 @@ const InlineInputs = ({
 
     return (
       <Label
-        labelId={labelId.current}
+        labelId={labelId}
         inline={inlineLabel}
         htmlFor={htmlFor}
         isRequired={required}
@@ -117,7 +107,7 @@ const InlineInputs = ({
         inputWidth={inputWidth}
       >
         <FormSpacingProvider marginBottom={undefined}>
-          {columnWrapper(children, gutter, label ? labelId.current : undefined)}
+          {columnWrapper(children, gutter)}
         </FormSpacingProvider>
       </StyledContentContainer>
     </StyledInlineInputs>